updated cloudflare page rules controller

+++ lib/modules/cloudflare/page_rule.rb
@@ -1,4 +1,4 @@
-module CloudFlare
+module Cloudflare
   require 'rubyflare'
 
   class PageRule
@@ -11,50 +11,62 @@ module CloudFlare
     # @param app_url [String] the application url
     # @return [String] returns the rule_id as a string
     def find_rule(domain, app_url)
-      client, zone_id = CloudFlare::Common.connect(domain)
+      client, zone_id = Cloudflare::Common.connect(domain)
       rules = client.get( "zones/#{zone_id}/pagerules")
       rule_id = nil
       rules.results.each do |result|
-        if result[:targets].first[:constraint][:value] == "#{app_url}/*"
+        if result[:targets].first[:constraint][:value] == "#{app_url}/*" && result[:actions].first[:value][:status_code] == 302
           rule_id = result[:id]
         end
       end
       return rule_id
     end
 
-    def create_rule(domain, app_url, forwarding_url)
-      client, zone_id = CloudFlare::Common.connect(domain)
-      redirect_url_rule ={
-                            targets: [
-                              {
-                                target: "url",
-                                constraint: {
-                                  operator: "matches",
-                                  value: "#{app_url}/*"
+    def create_redirect(domain, app_url, forwarding_url)
+      begin
+        client, zone_id = Cloudflare::Common.connect(domain)
+        redirect_url_rule ={
+                              targets: [
+                                {
+                                  target: "url",
+                                  constraint: {
+                                    operator: "matches",
+                                    value: "#{app_url}/*"
+                                  }
                                 }
-                              }
-                            ],
-                            actions: [
-                              {
-                                id: "forwarding_url",
-                                value: {
-                                  url: forwarding_url,
-                                  status_code: 302
+                              ],
+                              actions: [
+                                {
+                                  id: "forwarding_url",
+                                  value: {
+                                    url: forwarding_url,
+                                    status_code: 302
+                                  }
                                 }
-                              }
-                            ],
-                            status: "active"
-                          }
-      rule = client.post( "zones/#{zone_id}/pagerules",
-                              redirect_url_rule
-                            )
-      return rule
+                              ],
+                              status: "active"
+                            }
+        rule = client.post( "zones/#{zone_id}/pagerules",
+                            redirect_url_rule
+                          )
+        return [201, { message: "Temporary redirect created on #{app_url} to #{forwarding_url}"}.to_json]
+      rescue => e
+        puts "An error occured."
+        [404, { message: "#{e.response.errors} - #{e.response.messages}" }.to_json ]
+      end
+
     end
 
-    def delete_rule(domain, rule_id)
-      client, zone_id = CloudFlare::Common.connect(domain)
-      rule = client.delete("zones/#{zone_id}/pagerules/#{rule_id}")
-      return rule
+    def delete_rule(domain, app_url)
+      client, zone_id = Cloudflare::Common.connect(domain)
+      rule_id = find_rule(domain, app_url)
+      # binding.pry
+      if rule_id
+        rule = client.delete("zones/#{zone_id}/pagerules/#{rule_id}")
+        [201, message: "Page Rule deleted for #{app_url}".to_json]
+      else
+        [404, message: "Page Rule not found for #{app_url}".to_json]
+      end
     end
   end
 end
